requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic==2.10.6
requests==2.31.0
openai==1.30.1
pyyaml==6.0.1